+++ src/voxel/terrain_renderer.rs
@@ -2,186 +2,9 @@ use std::{sync::Arc, cell::RefCell};
 
 use crate::{math::Vec3, rendering::{VertexData, VertexBuffer, construct_render_pipeline, RenderPipelineInfo, DrawCall, IVertexBuffer, RenderStage}, camera::{Camera, CameraUniform}};
 use crate::gpu_utils::{Storage, BindGroup, Uniform, GPUVec3, Entry};
+use crate::voxel::voxel_rendering::*;
 
-pub enum FaceDir
-{
-    Up,
-    Down,
-    North,
-    South,
-    East,
-    West
-}
-
-impl FaceDir
-{
-    pub fn to_index(&self) -> u32
-    {
-        match self 
-        {
-            FaceDir::Up =>      0,
-            FaceDir::Down =>    1,
-            FaceDir::North =>   2,
-            FaceDir::South =>   3,
-            FaceDir::East =>    4,
-            FaceDir::West =>    5,
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct VoxelFace
-{
-    pos_x: u32,
-    pos_y: u32,
-    pos_z: u32,
-    direction: u32,
-    voxel_id: u32,
-}
-
-unsafe impl bytemuck::Pod for VoxelFace {}
-unsafe impl bytemuck::Zeroable for VoxelFace {}
-
-impl VoxelFace
-{
-    pub fn new(location: Vec3<u32>, direction: FaceDir, voxel_id: u16) -> Self 
-    {
-        Self 
-        { 
-            pos_x: location.x, 
-            pos_y: location.y, 
-            pos_z: location.z, 
-            direction: direction.to_index(), 
-            voxel_id: voxel_id as u32 
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct VoxelVertex
-{
-    pub face_index: u32,
-}
-
-unsafe impl bytemuck::Pod for VoxelVertex {}
-unsafe impl bytemuck::Zeroable for VoxelVertex {}
-
-impl VoxelVertex
-{
-    pub fn new(face_index: u32) -> Self 
-    {
-        Self 
-        { 
-            face_index, 
-        }
-    }
-}
-
-impl VertexData for VoxelVertex
-{
-    fn desc() -> wgpu::VertexBufferLayout<'static>
-    {
-        const ATTRIBUTES: [wgpu::VertexAttribute; 1] =
-            wgpu::vertex_attr_array![0 => Uint32];
-
-        wgpu::VertexBufferLayout {
-            array_stride: std::mem::size_of::<Self>() as wgpu::BufferAddress,
-            step_mode: wgpu::VertexStepMode::Vertex,
-            attributes: &ATTRIBUTES,
-        }
-    }
-
-    fn append_bytes(&self, bytes: &mut Vec<u8>) 
-    {
-        bytes.extend(bytemuck::cast_slice(&[*self]));
-    }
-}
-
-pub struct VoxelMesh
-{
-    vertices: Vec<VoxelVertex>,
-    faces: Vec<VoxelFace>,
-}
-
-impl VoxelMesh
-{
-    pub fn faces(&self) -> &Vec<VoxelFace> { &self.faces }
-    pub fn vertices(&self) -> &Vec<VoxelVertex> { &self.vertices }
-
-    pub fn new() -> Self 
-    {
-        Self 
-        { 
-            vertices: vec![], 
-            faces: vec![],
-        }
-    }
-
-    pub fn add_face(&mut self, location: Vec3<u32>, direction: FaceDir, voxel_id: u16)
-    {
-        let face_count = self.faces.len() as u32;
-        self.vertices.extend(
-        [
-            VoxelVertex::new(face_count),
-            VoxelVertex::new(face_count),
-            VoxelVertex::new(face_count),
-            VoxelVertex::new(face_count),
-            VoxelVertex::new(face_count),
-            VoxelVertex::new(face_count)
-        ]);
-
-        self.faces.push(VoxelFace::new(location, direction, voxel_id))
-    }
-
-    pub fn to_render_data(&self, voxel_size: f32, position: Vec3<u32>, device: &wgpu::Device) -> VoxelRenderData
-    {
-        let vertex_buffer = VertexBuffer::new(&self.vertices, device, Some("Voxel Vertex Buffer"));
-        let face_storage = Storage::new(&self.faces, wgpu::ShaderStages::VERTEX, device);
-        let voxel_size_uniform = Uniform::new(voxel_size, wgpu::ShaderStages::VERTEX, device);
-        let position_uniform = Uniform::new(position.into(), wgpu::ShaderStages::VERTEX, device);
-
-        VoxelRenderData::new(vertex_buffer, face_storage, voxel_size_uniform, position_uniform, device)
-    }
-}
-
-pub struct VoxelRenderData
-{
-    vertex_buffer: VertexBuffer<VoxelVertex>,
-    face_storage: Storage<VoxelFace>,
-    voxel_size_uniform: Uniform<f32>,
-    position_uniform: Uniform<GPUVec3<u32>>,
-    
-    bind_group: BindGroup
-}
-
-impl VoxelRenderData
-{
-    pub fn new(vertex_buffer: VertexBuffer<VoxelVertex>, face_storage: Storage<VoxelFace>, voxel_size_uniform: Uniform<f32>, position_uniform: Uniform<GPUVec3<u32>>, device: &wgpu::Device) -> Self
-    {
-        let bind_group = BindGroup::new(&[&face_storage, &voxel_size_uniform, &position_uniform], device);
-
-        Self 
-        { 
-            vertex_buffer, 
-            face_storage,
-            voxel_size_uniform,
-            position_uniform,
-            bind_group
-        }
-    }
-
-    pub fn construct_layout(device: &wgpu::Device) -> wgpu::BindGroupLayout
-    {
-        let face_storage_layout = Storage::<VoxelFace>::get_layout_static(wgpu::ShaderStages::VERTEX, 0);
-        let voxel_size_layout = Uniform::<f32>::get_layout_static(wgpu::ShaderStages::VERTEX, 1);
-        let position_layout = Uniform::<GPUVec3<u32>>::get_layout_static(wgpu::ShaderStages::VERTEX, 2);
-        BindGroup::construct_layout_from_entries(&[face_storage_layout, voxel_size_layout, position_layout], device)
-    }
-}
-
-pub struct VoxelRenderStage
+pub struct VoxelTerrainRenderStage
 {
     device: Arc<wgpu::Device>,
     
@@ -194,7 +17,7 @@ pub struct VoxelRenderStage
     render_pipeline: wgpu::RenderPipeline,
 }
 
-impl VoxelRenderStage
+impl VoxelTerrainRenderStage
 {
     pub fn new(camera: Camera, device: Arc<wgpu::Device>, config: &wgpu::SurfaceConfiguration) -> Self 
     {
@@ -238,7 +61,7 @@ impl VoxelRenderStage
     }
 }
 
-impl RenderStage for VoxelRenderStage
+impl RenderStage for VoxelTerrainRenderStage
 {
     fn render_pipeline(&self) -> &wgpu::RenderPipeline 
     {
@@ -279,7 +102,7 @@ impl<'a> DrawCall for VoxelDrawCall<'a>
 {
     fn bind_groups(&self) -> Box<[&BindGroup]> 
     {
-        Box::new([&self.camera_bind_group, &self.render_data.bind_group])
+        Box::new([&self.camera_bind_group, &self.render_data.bind_group()])
     }
 
     fn on_pre_draw(&self, queue: &wgpu::Queue) 
@@ -291,7 +114,7 @@ impl<'a> DrawCall for VoxelDrawCall<'a>
 
     fn on_draw<'pass, 's: 'pass>(&'s self, render_pass: &mut wgpu::RenderPass<'pass>) 
     {
-        render_pass.set_vertex_buffer(0, self.render_data.vertex_buffer.slice_all());
-        render_pass.draw(0..(self.render_data.vertex_buffer.capacity() as u32), 0..1);
+        render_pass.set_vertex_buffer(0, self.render_data.vertex_buffer().slice_all());
+        render_pass.draw(0..(self.render_data.vertex_buffer().capacity() as u32), 0..1);
     }
 }
